pyproject.toml
[project]
name = "r3lay"
version = "0.2.0"
description = "TUI research assistant with local LLM, hybrid RAG, and deep research capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.47.0",
    "httpx>=0.27.0",
    "ruamel.yaml>=0.18.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "rank-bm25>=0.2.2",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
rerank = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
]

[project.scripts]
r3lay = "r3lay.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
